ProjectMgtService-GettingStarted: Fixing some relationships between business model.
