+++ src/main/java/com/umssonline/proymgt/models/Task.java
@@ -5,13 +5,18 @@ import javax.persistence.*;
 @Entity
 public class Task extends SprintItem {
 
+    //region Properties
     @ManyToOne(optional = false)
     private UserStory userStory;
+    //endregion
 
+    //region Constructors
     protected Task() {
 
     }
+    //endregion
 
+    //region Getters & Setters
     public UserStory getUserStory() {
         return userStory;
     }
@@ -19,4 +24,5 @@ public class Task extends SprintItem {
     public void setUserStory(UserStory userStory) {
         this.userStory = userStory;
     }
+    //endregion
 }
